IN-21 change button storyes add linkAsButton add className of button
